requirements.txt
selenium==4.16.0
requests==2.31.0
Pillow==10.1.0
reportlab==4.0.8
beautifulsoup4==4.12.2
lxml==4.9.3
PyPDF2==3.0.1
webdriver-manager==4.0.1